+++ apps/ddaas/src/components/mdx-components.tsx
@@ -1,12 +1,8 @@
 import { createSiteMdxComponents } from '@windrun-huaiin/third-ui/fuma/server/site-mdx/base';
-// code
-// import { createCodeMdxComponents } from '@windrun-huaiin/third-ui/fuma/server/site-mdx/features/code';
-// math
-// import { createMathMdxComponents } from '@windrun-huaiin/third-ui/fuma/server/site-mdx/features/math';
-// mermaid
-// import { createMermaidMdxComponents } from '@windrun-huaiin/third-ui/fuma/server/site-mdx/features/mermaid';
-// type-table
-// import { createTypeTableMdxComponents } from '@windrun-huaiin/third-ui/fuma/server/site-mdx/features/type-table';
+import { createCodeMdxComponents } from '@windrun-huaiin/third-ui/fuma/server/site-mdx/features/code';
+import { createMathMdxComponents } from '@windrun-huaiin/third-ui/fuma/server/site-mdx/features/math';
+import { createMermaidMdxComponents } from '@windrun-huaiin/third-ui/fuma/server/site-mdx/features/mermaid';
+import { createTypeTableMdxComponents } from '@windrun-huaiin/third-ui/fuma/server/site-mdx/features/type-table';
 import { globalLucideIcons } from '@base-ui/icons';
 import { appConfig } from '@/lib/appConfig';
 
@@ -17,18 +13,18 @@ export const getMDXComponents = createSiteMdxComponents({
   },
   features: [
     // code
-    // createCodeMdxComponents(globalLucideIcons),
+    createCodeMdxComponents(),
 
     // math
-    // createMathMdxComponents(),
+    createMathMdxComponents(),
 
     // npm, no need components render, just need source handler
 
     // mermaid
-    // createMermaidMdxComponents(appConfig.style.watermark.enabled, appConfig.style.watermark.text),
+    createMermaidMdxComponents(appConfig.style.watermark.enabled, appConfig.style.watermark.text),
 
     // type-table
-    // createTypeTableMdxComponents(),
+    createTypeTableMdxComponents(),
   ],
   additionalComponents: {
     ...globalLucideIcons,