+++ packages/third-ui/src/fuma/server/optional-features.tsx
@@ -1,6 +1,0 @@
-export { createBaseMdxComponents } from './features/base';
-export { createCodeMdxComponents } from './features/code';
-export { createMathMdxComponents } from './features/math';
-export { createMermaidMdxComponents } from './features/mermaid';
-export { createTypeTableMdxComponents } from './features/type-table';
-export { createWidgetMdxComponents } from './features/widgets';