+++ apps/ddaas/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @windrun-huaiin/ddaas-website
 
+## 23.1.0
+
+### Minor Changes
+
+- fix(mdx): support cut packages by adjust mdx config
+
+### Patch Changes
+
+- Updated dependencies
+  - @windrun-huaiin/lib@23.1.0
+  - @windrun-huaiin/contracts@23.1.0
+  - @windrun-huaiin/base-ui@23.1.0
+  - @windrun-huaiin/third-ui@23.1.0
+  - @windrun-huaiin/fumadocs-local-md@23.1.0
+
 ## 23.0.0
 
 ### Major Changes